+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow customizing maximum retry delay in WriteToPostgres.
+
 ## [0.2.3] - 2022-11-24
 
 ### Fixed